# dwes06/dwes06Servicio/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Talleres y Ubicaciones",
        "description": "Listado de URI's de la API de Talleres y Ubicaciones",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        }
    ],
    "paths": {
...
        "/api/ubicaciones/{idUbicacion}/talleres": {
            "get": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Devuelve un listado con los talleres de una ubicación.",
                "operationId": "678e86406a73176ed5bc8025cce821a6",
                "parameters": [
                    {
                        "name": "idUbicacion",
                        "in": "path",
                        "description": "Identificador de la ubicación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "ubicacion_id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Taller de lectura"
                                            },
                                            "descripcion": {
                                                "type": "string",
                                                "example": "Taller de lectura en la biblioteca"
                                            },
                                            "dia_semana": {
                                                "type": "string",
                                                "example": "L"
                                            },
                                            "hora_inicio": {
                                                "type": "string",
                                                "example": "10:00:00"
                                            },
                                            "hora_fin": {
                                                "type": "string",
                                                "example": "11:00:00"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2023-02-23T00:09:16.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "example": "2023-02-23T12:33:45.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ubicación no existente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}